rename relay message to be `WalletLinkEventData`
