+++ packages/wallet-sdk/src/connection/WalletLinkConnection.ts
@@ -1,8 +1,8 @@
 // Copyright (c) 2018-2023 Coinbase, Inc. <https://www.coinbase.com/>
 // Licensed under the Apache License, version 2.0
 
-import { RelayMessage, Web3ResponseMessage } from '../relay/RelayMessage';
 import { Session } from '../relay/Session';
+import { WalletLinkEventData, Web3ResponseMessage } from '../relay/WalletLinkEvent';
 import { APP_VERSION_KEY, WALLET_USER_NAME_KEY } from '../relay/WalletSDKRelayAbstract';
 import { IntNumber } from '../types';
 import { ClientMessage } from './ClientMessage';
@@ -348,7 +348,11 @@ export class WalletLinkConnection {
    * @param callWebhook whether the webhook should be invoked
    * @returns a Promise that emits event ID when successful
    */
-  public async publishEvent(event: string, unencryptedMessage: RelayMessage, callWebhook = false) {
+  public async publishEvent(
+    event: string,
+    unencryptedMessage: WalletLinkEventData,
+    callWebhook = false
+  ) {
     const data = await this.cipher.encrypt(
       JSON.stringify({
         ...unencryptedMessage,
